refactor: 상세 모달의 버튼 분리
