+++ src/entity/comment/api/deleteComment.ts
@@ -1,0 +1,6 @@
+export const deleteComment = async (commentId: number) => {
+  const response = await fetch(`/api/comments/${commentId}`, {
+    method: "DELETE",
+  })
+  return await response.json()
+}
