+++ src/features/comment/add-button/index.tsx
@@ -1,0 +1,17 @@
+import { Button } from "@/shared/ui"
+import { Plus } from "lucide-react"
+
+export const CommentAddButton = () => {
+  return (
+    <Button
+      size="sm"
+      // onClick={() => {
+      //   setNewComment((prev) => ({ ...prev, postId }))
+      //   setShowAddCommentDialog(true)
+      // }}
+    >
+      <Plus className="w-3 h-3 mr-1" />
+      댓글 추가
+    </Button>
+  )
+}
